+++ trend_micro/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2.0.0 (Required minimum CE version for this is 6.1.0)
+## Added
+- Added support for new data regions with custom region configuration option.
+- Added new configuration parameters Base URL for custom region, Enable Push Retraction and Retraction Interval (in days).
+- Added support for pull and push retraction.
+- Added resolution for error logs.
+## Fixed
+- Updated indicator type mapping to correctly distinguish IPv4, IPv6, and Domain.
+- Updated retry handling to rely on the Retry-After header.
+## Changed
+- Updated plugin logo.
+
 # 1.0.2
 ## Fixed
 - Fixed Bugs in Push Functionality.